Rename utils to offline_utils and add created_timestamp_column
