+++ sdk/python/feast/infra/offline_stores/redshift.py
@@ -12,7 +12,7 @@ from feast import RedshiftSource
 from feast.data_source import DataSource
 from feast.errors import InvalidEntityType
 from feast.feature_view import FeatureView
-from feast.infra.offline_stores import utils
+from feast.infra.offline_stores import offline_utils
 from feast.infra.offline_stores.offline_store import OfflineStore, RetrievalJob
 from feast.infra.utils import aws_utils
 from feast.registry import Registry
@@ -113,31 +113,31 @@ class RedshiftOfflineStore(OfflineStore):
         )
         s3_resource = aws_utils.get_s3_resource(config.offline_store.region)
 
-        table_name = utils.get_temp_entity_table_name()
+        table_name = offline_utils.get_temp_entity_table_name()
 
         entity_schema = _upload_entity_df_and_get_entity_schema(
             entity_df, redshift_client, config, s3_resource, table_name
         )
 
-        entity_df_event_timestamp_col = utils.infer_event_timestamp_from_entity_df(
+        entity_df_event_timestamp_col = offline_utils.infer_event_timestamp_from_entity_df(
             entity_schema
         )
 
-        expected_join_keys = utils.get_expected_join_keys(
+        expected_join_keys = offline_utils.get_expected_join_keys(
             project, feature_views, registry
         )
 
-        utils.assert_expected_columns_in_entity_df(
+        offline_utils.assert_expected_columns_in_entity_df(
             entity_schema, expected_join_keys, entity_df_event_timestamp_col
         )
 
         # Build a query context containing all information required to template the Redshift SQL query
-        query_context = utils.get_feature_view_query_context(
+        query_context = offline_utils.get_feature_view_query_context(
             feature_refs, feature_views, registry, project,
         )
 
         # Generate the Redshift SQL query from the query context
-        query = utils.build_point_in_time_query(
+        query = offline_utils.build_point_in_time_query(
             query_context,
             left_table_query_string=table_name,
             entity_df_event_timestamp_col=entity_df_event_timestamp_col,
